refactor: move provider listing to its service
refs #1272
