+++ apps/api/src/provider/routes/providers/providers.router.ts
@@ -1,8 +1,9 @@
-import { createRoute, z } from "@hono/zod-openapi";
+import { createRoute } from "@hono/zod-openapi";
+import { container } from "tsyringe";
 
-import { cacheKeys, cacheResponse } from "@src/caching/helpers";
 import { OpenApiHonoHandler } from "@src/core/services/open-api-hono-handler/open-api-hono-handler";
-import { getProviderList } from "@src/services/db/providerStatusService";
+import { ProviderController } from "@src/provider/controllers/provider/provider.controller";
+import { ProviderListQuerySchema, ProviderListResponseSchema } from "@src/provider/http-schemas/provider.schema";
 
 const route = createRoute({
   method: "get",
@@ -10,101 +11,14 @@ const route = createRoute({
   summary: "Get a list of providers.",
   tags: ["Providers"],
   request: {
-    query: z.object({
-      scope: z.enum(["all", "trial"]).default("all")
-    })
+    query: ProviderListQuerySchema
   },
   responses: {
     200: {
       description: "Returns a list of providers",
       content: {
         "application/json": {
-          schema: z.array(
-            z.object({
-              owner: z.string(),
-              name: z.string(),
-              hostUri: z.string(),
-              createdHeight: z.number(),
-              email: z.string().nullable(),
-              website: z.string().nullable(),
-              lastCheckDate: z.string().nullable(),
-              deploymentCount: z.number(),
-              leaseCount: z.number(),
-              cosmosSdkVersion: z.string(),
-              akashVersion: z.string(),
-              ipRegion: z.string().nullable(),
-              ipRegionCode: z.string().nullable(),
-              ipCountry: z.string().nullable(),
-              ipCountryCode: z.string().nullable(),
-              ipLat: z.string().nullable(),
-              ipLon: z.string().nullable(),
-              uptime1d: z.number(),
-              uptime7d: z.number(),
-              uptime30d: z.number(),
-              isValidVersion: z.boolean(),
-              isOnline: z.boolean(),
-              lastOnlineDate: z.string().nullable(),
-              isAudited: z.boolean(),
-              activeStats: z.object({
-                cpu: z.number(),
-                gpu: z.number(),
-                memory: z.number(),
-                storage: z.number()
-              }),
-              pendingStats: z.object({
-                cpu: z.number(),
-                gpu: z.number(),
-                memory: z.number(),
-                storage: z.number()
-              }),
-              availableStats: z.object({
-                cpu: z.number(),
-                gpu: z.number(),
-                memory: z.number(),
-                storage: z.number()
-              }),
-              gpuModels: z.array(
-                z.object({
-                  vendor: z.string(),
-                  model: z.string(),
-                  ram: z.string(),
-                  interface: z.string()
-                })
-              ),
-              attributes: z.array(
-                z.object({
-                  key: z.string(),
-                  value: z.string(),
-                  auditedBy: z.array(z.string())
-                })
-              ),
-              host: z.string().nullable(),
-              organization: z.string().nullable(),
-              statusPage: z.string().nullable(),
-              locationRegion: z.string().nullable(),
-              country: z.string().nullable(),
-              city: z.string().nullable(),
-              timezone: z.string().nullable(),
-              locationType: z.string().nullable(),
-              hostingProvider: z.string().nullable(),
-              hardwareCpu: z.string().nullable(),
-              hardwareCpuArch: z.string().nullable(),
-              hardwareGpuVendor: z.string().nullable(),
-              hardwareGpuModels: z.array(z.string()),
-              hardwareDisk: z.array(z.string()),
-              featPersistentStorage: z.boolean(),
-              featPersistentStorageType: z.array(z.string()),
-              hardwareMemory: z.string().nullable(),
-              networkProvider: z.string().nullable(),
-              networkSpeedDown: z.number(),
-              networkSpeedUp: z.number(),
-              tier: z.string().nullable(),
-              featEndpointCustomDomain: z.boolean(),
-              workloadSupportChia: z.boolean(),
-              workloadSupportChiaCapabilities: z.array(z.string()),
-              featEndpointIp: z.boolean()
-            })
-          )
+          schema: ProviderListResponseSchema
         }
       }
     }
@@ -115,13 +29,7 @@ export const providersRouter = new OpenApiHonoHandler();
 
 providersRouter.openapi(route, async function routeListProviders(c) {
   const { scope } = c.req.valid("query");
-
-  const providers = await cacheResponse(
-    60,
-    scope === "trial" ? cacheKeys.getTrialProviderList : cacheKeys.getProviderList,
-    () => getProviderList({ trial: scope === "trial" }),
-    true
-  );
+  const providers = await container.resolve(ProviderController).getProviderList(scope);
 
   return c.json(providers);
 });
